code/requirements.txt
flask
flask-RESTful
flask-JWT
flask-SQLAlchemy
uswgi